change file names to rodneyClicker (#17)
