+++ app/src/androidTest/java/com/example/rodneyClicker/SaveDataTest.kt
@@ -1,7 +1,7 @@
-package com.example.helloworld
+package com.example.rodneyClicker
 
 import androidx.test.espresso.Espresso.onView
-import androidx.test.espresso.action.ViewActions.click
+import androidx.test.espresso.action.ViewActions.doubleClick
 import androidx.test.espresso.action.ViewActions.replaceText
 import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.ViewMatchers.withId
@@ -13,12 +13,12 @@ import org.junit.Test
 import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
-class RavenButtonTest {
+class SaveDataTest {
     @get:Rule
     var rule = activityScenarioRule<MainActivity>()
 
     @Test
-    fun clickerIncrement() {
+    fun saveData() {
         // Set Raven Dollars to 0
         onView(withId(R.id.ravenDollars)).perform(replaceText("0"))
 
@@ -26,13 +26,16 @@ class RavenButtonTest {
         onView(withId(R.id.ravenDollars)).check(matches(withText("0")))
 
         // Click the button
-        onView(withId(R.id.ravenButton)).perform(click())
+        onView(withId(R.id.ravenButton)).perform(doubleClick())
+        onView(withId(R.id.ravenButton)).perform(doubleClick())
 
         // Confirm that value increments
-        onView(withId(R.id.ravenDollars)).check(matches(withText("1")))
+        onView(withId(R.id.ravenDollars)).check(matches(withText("4")))
+    }
 
-        // Click again and confirm new value
-        onView(withId(R.id.ravenButton)).perform(click())
-        onView(withId(R.id.ravenDollars)).check(matches(withText("2")))
+    @Test
+    fun loadData() {
+        // Check to see if the value is saved
+        onView(withId(R.id.ravenDollars)).check(matches(withText("4")))
     }
 }
