--- app/src/main/java/com/example/helloworld/DisplayMessageActivity.kt
@@ -1,20 +1,0 @@
-package com.example.helloworld
-
-import android.os.Bundle
-import android.provider.AlarmClock.EXTRA_MESSAGE
-import android.widget.TextView
-import androidx.appcompat.app.AppCompatActivity
-
-class DisplayMessageActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_display_message)
-        // Get the Intent that started this activity and extract the string
-        val message = intent.getStringExtra(EXTRA_MESSAGE)
-
-        // Capture the layout's TextView and set the string as its text
-        val textView = findViewById<TextView>(R.id.textView).apply {
-            text = message
-        }
-    }
-}
