+++ app/src/test/java/com/example/rodneyClicker/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.example.helloworld
+package com.example.rodneyClicker
 
 import org.junit.Assert.assertEquals
 import org.junit.Test
